requirements.txt
alpaca-trade-api==1.2.3
pyyaml==6.0